+++ init.rb
@@ -1,5 +1,9 @@
-# View hook
-require_dependency 'redmine_recurring_tasks/hooks/issue_sidebar_hook'
+ActionDispatch::Callbacks.to_prepare do
+  paths = '/lib/redmine_recurring_tasks/{patches/*_patch,hooks/*_hook}.rb'
+  Dir.glob(File.dirname(__FILE__) + paths).each do |file|
+    require_dependency file
+  end
+end
 
 Redmine::Plugin.register :redmine_recurring_tasks do
   name 'Redmine Recurring Tasks'