+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * **(breaking-change)** Rename plugin to RedmineRecurringTasks
 * **(breaking-change)** Transfer plugin to southbridge (http://github.com/centosadmin)
 * Make preserve author of issue on copy
+* Add relation between Issue and WeeklySchedule 
 
 # 0.0.1
 * Add ability to make issues copies